+++ client/src/game/Game.vue
@@ -85,14 +85,17 @@ export default class Game extends Vue {
     // Touch events
 
     touchend(event: TouchEvent): void {
+        if (this.$refs.ui === undefined) return;
         this.$refs.ui.$refs.tools.touchend(event);
     }
 
     touchstart(event: TouchEvent): void {
+        if (this.$refs.ui === undefined) return;
         this.$refs.ui.$refs.tools.touchstart(event);
     }
 
     touchmove(event: TouchEvent): void {
+        if (this.$refs.ui === undefined) return;
         // limit the number of touch moves to ease server load
         if (!this.throttledtouchmoveSet) {
             this.throttledtouchmoveSet = true;
@@ -105,14 +108,17 @@ export default class Game extends Vue {
     // Mouse events
 
     mousedown(event: MouseEvent): void {
+        if (this.$refs.ui === undefined) return;
         this.$refs.ui.$refs.tools.mousedown(event);
     }
 
     mouseup(event: MouseEvent): void {
+        if (this.$refs.ui === undefined) return;
         this.$refs.ui.$refs.tools.mouseup(event);
     }
 
     mousemove(event: MouseEvent): void {
+        if (this.$refs.ui === undefined) return;
         if (!this.throttledmoveSet) {
             this.throttledmoveSet = true;
             this.throttledmove = throttle(this.$refs.ui.$refs.tools.mousemove, 15);
@@ -121,10 +127,12 @@ export default class Game extends Vue {
     }
 
     mouseleave(event: MouseEvent): void {
+        if (this.$refs.ui === undefined) return;
         this.$refs.ui.$refs.tools.mouseleave(event);
     }
 
     contextmenu(event: MouseEvent): void {
+        if (this.$refs.ui === undefined) return;
         this.$refs.ui.$refs.tools.contextmenu(event);
     }
 
@@ -146,7 +154,7 @@ export default class Game extends Vue {
 
 <template>
     <div id="main" @mouseleave="mouseleave" @wheel="zoom">
-        <ui ref="ui"></ui>
+        <ui v-if="$store.state.game.boardInitialized" ref="ui"></ui>
         <div id="board">
             <div
                 id="layers"