+++ client/src/game/api/events.ts
@@ -6,7 +6,7 @@ import "@/game/api/events/location";
 import "@/game/api/events/room";
 import "@/game/api/events/shape";
 import { socket } from "@/game/api/socket";
-import { BoardInfo, Note } from "@/game/comm/types/general";
+import { BoardInfo, Note, ServerFloor } from "@/game/comm/types/general";
 import { EventBus } from "@/game/event-bus";
 import { GlobalPoint } from "@/game/geom";
 import { layerManager } from "@/game/layers/manager";
@@ -23,6 +23,7 @@ import { floorStore } from "../layers/store";
 
 socket.on("connect", () => {
     console.log("Connected");
+    socket.emit("Location.Load");
 });
 socket.on("disconnect", () => {
     console.log("Disconnected");
@@ -54,9 +55,7 @@ socket.on("Client.Options.Set", (options: ServerClient) => {
     gameStore.setPanY(options.pan_y);
     gameStore.setZoomDisplay(zoomDisplay(options.zoom_factor));
 
-    const floor = options.active_floor ?? 0;
-    socket.once("Board.Set", () => {
-        floorStore.selectFloor({ targetFloor: floor, sync: false });
+    socket.once("Board.Floor.Set", () => {
         if (options.active_layer) layerManager.selectLayer(options.active_layer, false);
     });
 });
@@ -76,6 +75,27 @@ socket.on("Board.Set", (locationInfo: BoardInfo) => {
     }
     requestAnimationFrame(layerManager.drawLoop);
     gameStore.setBoardInitialized(true);
+});
+
+socket.on("Board.Locations.Set", (locationInfo: { id: number; name: string }[]) => {
+    gameStore.clear();
+    visibilityStore.clear();
+    gameStore.setLocations({ locations: locationInfo, sync: false });
+    document.getElementById("layers")!.innerHTML = "";
+    floorStore.reset();
+    layerManager.reset();
+    EventBus.$emit("Initiative.Clear");
+});
+
+socket.on("Board.Floor.Set", (floor: ServerFloor) => {
+    addFloor(floor);
+    visibilityStore.recalculateVision(floor.name);
+    visibilityStore.recalculateMovement(floor.name);
+    if (floorStore.floors.length === 1) {
+        floorStore.selectFloor({ targetFloor: floor.name, sync: false });
+        requestAnimationFrame(layerManager.drawLoop);
+        gameStore.setBoardInitialized(true);
+    }
 });
 
 // Varia